improve reliability of spin tests
They fail often with Asan. That's because ASAN extends the execution time
of most things by quite a bit.

Even for the spin test, note that we should not verify that we ran for
the entire duration of the spin: if there are other threads running together
with us they would have gotten their share too: all the spin does is making
sure that a certain duration passes.