Use standard code-path for canceling timers
Previously, we canceled the timeout timer in a manual way, which
worked, but was a bit roundabout.

Specifically, there was a weird case when the timeout Source was
present in the SourceMap twice:

* once as an expired timer with a pending cancellation
* once as a freshly armed timer

With this PR though, we just allocate a fresh source when we want to
rearm the timer. This allows for intuitive "just drop the alarm on the
floor, hard" way of canceling the timer!