add comments about timers
Aleksey was already the second person to feel confused by our "two
timers".

This code evolved organically so the names are still a bit unfortunate.

This patch cleans it up, renames arm_preempt_timer to just arm_timer
(since that's now what it does) and hopefully explain in simple
language what is going on.