deadline queue: fix memory leak as items never leave the queue
When we admit an item we add it to a queue. That queue is used in the shares
calculation: we go over it and calculate the shares need based on all elements
that are currently being processed.

However we never pop it from this queue. The effect on the shares calculation
should be negligible: this will be a completed source, so it shouldn't
contribute much to the calculation.

But it does leak memory because the DeadlineSource is never destroyed.  I
didn't notice this even with Asan because whenever the entire DeadlineQueue is
destroyed, at the end of the test, our leaky DeadlineSource is destroyed too.

In this patch we fix this issue and add an explicit test to make sure that
we see the DeadlineSource destroyed when we expect it to be.