Dma writes are broken if file is bigger than write_behind * buffer_size
This is because we add elements to the pending list, but never take them
out. This code worked before because we kept a fixed size list of buffers
and gave buffers back. Pending had no role in determining whether or not
we could write.

The worst problem, though, was that the unit tests that should cover this
didn't: because the buffer size is aligned to the Direct I/O alignment
we don't really write with the buffer size we requested so all tests
ended up being too small to trigger this.

Fixes #141