improve controller memory
Because we are popping elements from a queue, it could be that between
one item finishing and the next starting we may end up with expected very
close to zero.

The current code brings the process to its zero state in that case - by
setting shares to one. But it would be better if we were to just return
last_shares in this case: whatever shares we converged to in the
last iteration of the process is likely to be a better starting point
than 1.