change calculation of active tasks
We determine whether or not a task is active based on whether we popped
the last element of the queue.

That worked well before we checked for preemption, because if we are processing
the last element of the queue now, we'll try to enter the loop once more and
conclude the tq is now inactive.

But with preemption we may end up popping the last element and breaking the loop
before we try again. In that case the task queue is inactive but we never marked
it as so.

We always call account_vruntime for every run so it is better to defer the
decision on whether to mark a task as inactive until that point.