executor: move available_executors to a BTreeMap
As Aleksey recently pointed out, HashMap in Rust has cryptographic strength
so it is a lot more expensive than what we need.

By switching to a BTreeMap we save around 100ns in the upcoming executor
benchmarks.

I have considered a FreeList here as well but the interface that we need
is a bit of a stretch: we need to store the index inside the item, and
also pre-allocate some indexes.

This is all doable with the freelist but I feel it starts to become a little
bit of an abuse of the API.