fix spurious wake ups.
We are waking up spuriously, because we only delete the old timer
in the following two situations:

1. when there is a preempt timer.
2. when there is a user timer

If we are just going to sleep without any timer the timer is not canceled.
Therefore, we can spuriously wake up if we go to sleep without any timer.

A better, safer way to write this is to always delete the timer:

Note that the arm paths are mutually exclusive, since if we are going to
sleep we can't preempt. So whichever path wants to register a timer can
do so.