sadly remove the restrictive unshare call
It is still a great idea maybe for safety, but it essentially means we
cannot use eventfds for inter-executor notification. unshare is unfortunately
a on/off switch so we can't unshare just some part of the fd space and not
others.

One of our tests relied on the behavior of close being safe because of unshare
so we'll have to remove it.