shared channels: spsc connection between executors
Sometimes we need to pass data between executors. The shared channel
will be used for that.

The idea behind the shared channel is that the types going through the channel
need to be Send. But in order to probe for whether or not there is space in the
channel and activate the wakers we need to drill down the reactor and check on
every poll. Because of that, we require a sender or receiver to be connected
with a connect() function.

The connect() function consumes self, the inner state is not accessible to the
public, and before a connection is made there are no methods available in
either the sender or receiver: this makes it safe to force the Send trait onto
the Shared Receiver/Producer even though their state is implemented through an
Rc.

The Rc is important for two reasons: first it allow us to pass the state deep
down to the reactor so we can check if the wakers can be woken up.

But secondly, it allows for many asynchronous, detached tasks to push data into
this channel inside a single executor.

To implement the Shared Channel I am using a modified version of the spsc queue
crate from https://github.com/polyfractal/bounded-spsc-queue. I like the
implementation and it is plenty fast but aside from small changes (that I tried
contributing back), for maximum performance I want to pass the eventfd
information as well as connectivity in the same cachelines as the
producer/consumer. So the changes became too big and I just copied the code.

The eventfd mechanism is needed so that one executor can wake up the other
in case it goes to sleep: before going to sleep the executor will post
its eventfd file descriptor into the ring (upon seeing that, the remote
executor will notify) and when it wakes up it sets it back to zero,
meaning it is not sleeping.

There is a benchmark that comes attached to it, with a similar code
for rust's mpsc channel implemented for comparison:

Using my local server, I see:

     Running target/release/deps/shared_channel-4e66644014f75ba1
cost of sending rust standard channel 174ns, capacity: 1024
cost of receiving rust standard channel: 173ns, capacity 1024
cost of sending shared channel 28ns, capacity: 1024
cost of receiving shared channel: 28ns, capacity 1024