implement poll_close for the buffered stream writer
We didn't implement poll_close, counting on the fact that our close
function would be called instead.

However, if we are calling close through a function that takes an
AsyncWriteExt trait the close method from the trait gets called. In that
case, we panic.

So let's go the extra mile and implement it: we'll need the file to
live inside an Option, so we can take() it. That is because the buffered
file's close() (now also accessible through discard()) consumes self.

We have a state machine similar to the one in the DMA file to handle the
transitions.

Once we have poll_close, then we no longer need our own close function.