simplify and coalesce io_uring dispatch
If we have cancellations we dispatch them right away with io_uring_enter,
only to call into the kernel again with the standard submissions soon.
This used to be necessary because we reused the id for the timer across
invocations, but now that we do not we can relax it and coalesce
cancellations and submissions.

The only requirement is still that cancellations manage to get all the
sqes it needs.

Once these are coalesced the uring code can be simplified as we have
fewer places now where the decision to sleep can be changed.