replace std hashmap with ahashmap
This is a non-cryptographic hash, therefore faster than rust's standard
hash.

Although we have a FreeList type, which is useful for simple integer
cases, the FreeList API is a bit more restrictive than the Hash and
converting some uses would require work.

Contrast this with AHash, which was essentially a bunch of sed invocations.

This gave me a 5% boost in the storage benchmark I am about to post.