Indicate whether a file supports Direct I/O in the PollableStatus
If the file does not support Direct I/O we don't want to use the
upcoming registered buffers support. This will be the case for tmpfs
files.

Following the principle of least surprise, we will also no longer
be disabling Direct I/O silently every time it fails. We will only
do that if we are dealing with a tmpfs file.