uring registered buffers
This patch introduces registered buffers for io_uring.
Registered buffers allows one to register a memory area with io_uring.
That area is pre-mapped and pinned, meaning that Direct I/O access
does not need to do that work for every request.

It is pointless to register one buffer per operation: instead we
register an entire area and use a buddy allocator in that area.

Buddy is the best algorithm in this case given the alignment guarantees
of Direct I/O, which pretty much means we'll be dealing with page-sized
blocks.

Not being able to register the buffers, however, is a non-fatal error.
Because this memory is accounted towards the user's RLIMIT_MEMLOCK we'll
allow applications to decrease the registration (or increase if you plan
to keep an insane amount of buffers in-flight).

Using up the entire area is non-fatal as well: we'll allocate memory
normally and incur a performance penalty.