Make the reactor a member of the Executor
The Reactor is currently initialized as a TLS value. This is because
at some point in the crates we used as base, those things could be
used independently. For us, they are not.

TLS in Rust is more expensive than we'd like. And with both the Reactor
and the Executor as TLS variables one has to at times use two TLS variables
to achieve a goal. With this design not only we reduce that to one, but it
becomes possible to add a Rc<Reactor> to long-lived data structures like
the various *File: we'll then access the reactor for reads and writes
with no TLS access at all.

More than that, it is hard to pass Executor parameters down the Reactor.
Now we'll be able to.

One concern is the cost of need_preempt(), which is called very often,
as scoped_thread_local is more expensive than the standard thread_local.

A couple of things are done to keep this in check. In particular we can
change the pattern of using if LOCAL_EX.is_set() { ... }: scoped_thread_local
will already panic if not set, so the extra check only give us a nicer
error message. Nice indeed, but not worth the cost. We apply that pattern
all over.

A benchmark is added so we can test the cost of need_preempt(). It does
get more expensive, but only by 1 ns.