rename the PosixDmaBuffer
We finally have an implementation of registered buffers! With this,
we need no longer imagine what the best way would be to manage various
kinds of buffers.

In the way this ended up, we always have the same external buffer
type and record internally for freeing purposes which buffer we are
dealing with.

Therefore it makes no sense to call this "PosixDmaBuffer" and we call
it a "DmaBuffer" instead (the fact that we had to export an alias for
"DmaBuffer" as a public type was a good indication that this was a
likely direction)

Now that the type is public, we'll be forced to write documentation
and massage the code to make clippy happy.

But that's a good thing!

Enjoy.