be nicer to our new friend, the borrow clippy lint
Rust 1.49 will include a new linter contributed by Daniel. It is already
present in the beta.

The new clippy lint step checks if borrows are held across suspend
points and fails if they are.

Our usage is currently safe, but the lint has issues with explicit
calls to drop and I was told by Daniel this can be challenging to fix.

Let's play nice and use explicit scopes to handle that.
After this patch we're all good according to 1.49's clippy.