provide a new method, from_rc for the DmaStreamReader
It is usually the case that a file opened for reads can be accessed
both linearly and randomly depending on the circumnstance. And random
access, as they are usually done through spawned tasks, would be using
an Rc already.

I considered just changing new() to accept and Rc and not having two methods.
But for the simple case where you only have linear access it feels like more
ergonomic not to worry about the fact that we use Rc internally. It also
would be inconsistent with the rest of the APIs.