fix hang with reading post explicit end of file
If we set an early-eof explicitly we loop forever trying to replenish
buffers. That's because we leave fill_buffer early, and never bump
neither resident buffers nor pending.

Try a fixed amount of times and bail.

There is no need of a test for this because one of our tests actually
fail as soon as the default read ahead is bumped, which is happening
in the next patch.