reduce visibility of Runnable and other multitask entities
While reviewing Andrey's code that touched the task core, I noticed
that Runnable is fully public, and that should not be the case.

As a matter of fact, nothing in the multitask file should be public,
since it only exists to implement an internal queue of the main
executor.

Reduce the visibility to all currently public types to pub(crate).