avoid allocation for time registration
The idea of always using a new source for the timer and relying on the old
one being dropped for cancellation is neat, but Sources as they are right now
must live in the heap.

That means that this code path always triggers a memory allocation. It is
best to cancel explicitly (we were already dropping explicitly, so that is
easy), and use the same source all the time, just updating its SourceType.