async close files on drop
As suggested on Zulip by user Gardner Vickers, since we now embed the reactor
in the File, we can safely asynchronously close.

The disadvantage is that the file going out of scope won't reduce the file
descriptor count, but it is guaranteed to register for the next poll so it is
not that bad. The file descriptor cannot be reused since a new open() will not
be able to acquire it, so this is safe.

Expliclit closing is still preferred, and a message about that is still
printed in debug level.