LocalExecutor::available_executors is converted from AHashMap to Vec.
Checking the id generation of task queues I found out that ids of the queues:
1. Do not need to be unique overall time of the application.
2. Managed internally.

That means that Map could be replaced by Vec<Option> with VecDeque which contains collection of ids of removed queues.
There is could be situation when queues will be massively created and then removed.
Which will create list of "null" references to the queues which only consumes memory,
but the same is true for the Map, also such situation when such massive spike of removal
of queues is hardly possible on practice.