Fix of issue #194. All changes in task status are guarded by the check of thread_id.
I have checked two implementations:
1. When Wakers are protected by executor_id.
2. When Wakers are protected by thread_id.

The last appeared to be more lightweight and personally I think that it is less fragile from the point of view of usage of API.
For both cases I did not notice noticeable slowdown in execution of benchmarks.

To improve performance thread_id detection is wrapped in call of thread_local. Even if now there are some questions
to its performance there is on going issue which should solve them.