specialized poll io loop that doesn't touch timer registration
Everybody has a routine before going to bed and we are no different.
Before sleeping, we may, if the user so desires, poll a bit before
declaring there isn't really any work and going down.

However because we have been using the standard poll loop for that,
every time we did that we had to cancel the existing preempt timer,
install a new one, etc.

The TcpStreams patchset brought in a rush_dispatch mechanism that we can
extend for the pre-sleep pollers so we'll take advantage of that

Fixes #76