Addresses #180: Implement error system
Uses 'thiserror' and some refactoring to implement a cohesive error type that
can encompass all errors produced by glommio. Includes all the natural traits
like from/into so try ('?') should work as expected if you are using it on the
inner type of whichever variant you are using. Also included in this commit is
an example migration of the semaphore errors as an example of how it would
likely work. Assuming this is satisfactory, the rest of the errors can be
refactored to use this as well. This means we shouldn't need to piggyback on the
std::io::Error to express our intent. It should still be used as the inner type
if underlying libraries produce this, but otherwise the correct variant of
GlommioError should just be manually constructed.