extract some tcp code to common code
This is an attempt to make the implementation of Udp / UnixSockets easier
or at least more elegant.

A lot of the code we have is just handling return conditions, mapping between
options, results and Polls, and fallbacks from the Yolo path.

With some helper macros and the yolo path extracted, implementing other
protocols should be simpler now.