migrate result type, consolidate sub-errors and add testing
Consolidated the error types that made sense to be embedded directly in the
error enum. Migrated the public facing results to use the new GlommioError as
well as a Result type definition (to remove the generic, when it was pointless).
Also added more tests for going from base errors -> GlommioError -> io::Error's.
Also added #[should_panic] tests that test the error messages a little more
directly.