fix Tcp bind
Discovered by Bryan.
The problem is that bind currently doesn't actually uphold ReusePort:
we bind first, and only then set ReusePort, while we should do the opposite.

The test is broken, because in some (most?) runs the executors will end up
serializing. That is because all tests run in parallel and turn all CPUs
busy.

The test is patched to make sure we are keeping the executors alive,
therefore keeping the socket alive too, until the other executor tries
to bind. Before the fix, the test now fails every time. After the fix,
the test no longer fails.