no longer use an std::file internally
We did that to force a close if the user didn't close().await.
However that forces a mem::forget() -> in case close().await is called.
That mem::forget() causes the destructor for the weak reference on the
reactor not to be called, and then asan sees that as a memory leak.

Ever since we started doing async close we no longer have to use an
std::file. We should as a matter of fact have cleaned that up back then.
With this patch, asan is happy again.

Fixes #229