Rc containers were removed from LocalChannel implementation.
Because Rc is a boxed type, Rc containers which were used inside of
LocalChannel implementation of waiting queue
were replaced by NonNull-s of Pinned waiting nodes.
To ensure safety the following changes were done:
1. Futures lifetimes were restricted by lifetime of the LocalChannel.
 That is quite minor restriction because
that is typically how sync primitives are used anyway.
2. All waiting nodes were made Pinned.
3. On Drop of the LocalChannel assertion that waiting queues are empty
was added.
4. On Drop Futures remove itself from waiting queue.
5. LocalChannel does not implement Stream itself but returns related Stream
instance through LocalReceiver#stream() method.
That makes usage less convenient but a) It is more canonical because single
Stream should work with with single Task instance the same as Future.
b) It makes implementation simpler.
6. On Drop of stream associated waiting node is removed from the waiting queue.