make every future go through a task queue.
We currently have a weird leftover requirement that the first future
that we pass to run() be special. It doesn't run in a task queue, because
that is the future we run to completion.

Such requirement creates an asymmetry that was sooner or later bound to
cause issues, as there are many things that you cannot do if you don't
have a task queue where you are running.

This patch moves the initial future inside a task queue as well. Now
everything is symmetric and there are no special cases.

Our spawn functions all force futures to be 'static, so we need to
remove this requirement. We keep the requirement in outter functions
because the ability to .detach() them means that we truly cannot
control the lifetime.

Fixes #241