Issue #229 has been fixed.  (#258)
* Issue #229 has been fixed. An issue that caused program abort, in the presence of not finished timer in the executor,  has been fixed.

Issue #229 has been caused by the fact that the timer kept instance of Waker, which was not used to wake the future. That drove cyclic reference between the not finished timer and future itself through waker instance.

Program abort was caused by the drop of the future that triggered drop of the waker.  As a result, waker tries to reschedule the local executor task, which causes panic because local executor is absent. The panic inside of local executor triggers drop of the timer, but the reactor is already absent at this moment which triggers panic inside panic and as a result triggers abort of the program.

Fixes #229