shared channels: make connect() return a future
It will unfortunately not be possible to fix #261 without an API change.
This is because the channel needs to keep the remote end - or at least
part of its state - alive to avoid a data race.

We don't know when - or if - the remote end will connect. Therefore this
needs to be made into a future that can resolve when the connection happens
(or when the endpoint disconnects)

This will be a bit complex, so it pays to add the API change separately
to reduce the change.

Note that some of the examples as-is won't work anymore: that is because
they were connecting both ends in the same thread. One can make that work
by spawning tasks, but the whole point of shared channels is to connect
different executors so the examples are changed to reflect that.