sleep notifier: a stronger wrapper around the eventfd
Issue #261 exposed a serious weakness in how we're handling remote
notifications: the lifetime of the eventfd and the memory area for
notifications is disconnected from each other, plus the remote end
of the channel has no way of keeping that alive.

To pave the way for solving this problem, this patch introduces the
SleepNotifier. Wrapped around an Arc, it will ease the task of allowing
the peer keeping the channel alive. Plus the lifetimes of all entities
involved in the notification is kept together.