shared channels: use the sleep notifiers to make channels safe.
When an executor goes to sleep, it's write into a memory area that is
stored in the shared buffer to signal to the other side of the shared
channel that it needs to be notified and wake up.

This works well with long-lived executors for which the shutdown process
is well behaved so that the channels will be always empty. While that is
still a good practice (we are adding that to the docs!) we don't want to
force that or depend on that for correctness.

In this patch we create a global table that maps executor IDs to their
SleepNotifier. Because the SleepNotifier is wrapped in an Arc, a connected
channel can keep its peer alive by holding onto that Arc. Even if the reactor
dies, the memory will still be valid and so will the eventfd.

Fixes #261