special-case the disconnected SleepNotifier
Because we currently don't do that, we can't use a Weak pointer to store
the sleep notifier. So when all references are dropped, we still keep a reference
in the map itself and the sleep notifier is never dropped.

Consequently, the eventfd file is never freed, which is a leak.

There are two ways to fix this:
- special case the disconnected notifier. The map now can contain weak
  references, and the notifier will be dropped when both ends disconnect
- get the owning reactor to manually remove the entry when it goes away.

Both works, but the former is more RAII-friendly so we go with that.