fix hangs with sleeping queue
There are some problems in our current spsc implementation that are related to
memory order. Those problems are fixed (in particular a Relaxed load from the
real tail), but none of them are really the root cause of what we are seeing.

The real root cause is that because of the possibility of reordering at the CPU
level, it could be that we marked ourselves as going to sleep and checked if new
work has happened, but our peer did those things in the opposite order.

A barrier is needed, that we were missing. Barriers are needed both in the producer
and the receiver side, which can be quite expensive. To avoid that, we are using
a technique much like the one described at
https://www.scylladb.com/2018/02/15/memory-barriers-seastar-linux/

Fully Fixes #261