spsc: don't do wrapping add
Wrapping add is more expensive than a normal add but this is pointless
because if we do wrap up, our code will break in other places prior
to that. For instance, the way size() is calculated is tail - head,
which is fast, but won't work with wrapping.

We are working with usize which is 64-bit in most modern platforms so
that should be enough. And when it is not, we'll panic (now with a test
to make sure of that)