add a version of the semaphore permit for complex lifetimes
Sometimes the lifetimes associated with the permits are complex. We may
move them, pass around, and keep them alive to signal that a long-lived
operation is still in progress.

In such sitations we need a more complex permit, that is able to keep
the semaphore alive for the lifetime of the operation.

In reality I expect most permits to be like this. But recognizing that
sometimes simple is good, I am adding a separate permit acquisition
for these cases instead of requiring that permits are always used with
an Rc.