special case the non-crossing case when skipping bytes
Although the code is more elegant the way it is right now, we need 3
for loops that will return immediately, meaning 3 branches. We can
add a single branch to the beginning which will help us skip them all.

This helps a lot the case of small reads where crossing buffers is rare.