shared_channels: clear wakers and connection maps when unregistering
When shared channels are destroyed, it could be that there are still
pending wakers. If they are destroyed there isn't much we should do,
as there is nobody else to be awaken. We clear the collections to avoid
letting other parts of the code thinking there are still notifications
to be sent.

Fixes #269