fix shared channels lock up
The previous commit ended up introducing a bug: we don't have to
flush the entire wakers list plus connections because they are global.
It caused unrelated shared channels to miss wake ups.

Only the waker to this id should be removed.
Also the connection vector is fine. It doesn't hurt if we have a
spurious wake up.