handle EBUSY returns from io_uring
When the CQE queue is empty, io_uring_enter can return EBUSY, essentially
blocking the submission of new requests. This is non-fatal: this means
that the system is overloaded, and we just have to keep queueing
requests until we are finally able to dispatch.

This is especially important for rushed mode, where we assume that
a response may be available immediately after dispatch and speculatively
read more: if we can't dispatch right away, just do it later.