WIP: fix error messages data loss
Removes thiserror b/c we are forced to implement the display impl by hand,
leaving the crate pointless for our purposes. Our needs were just too
complicated, for this crate to be used. Instead we can now actually give custom
messages for WouldBlock and Closed messages based on the resource type. Because
some include space for extra fields, or messages.