Additional tests and tweaks for non-copy shared channel
Added many more tests to verify drop behavior, and attempted to simulate all
possible scenarios. Also managed to remove both the Option, and an inner Arc
that didn't need to exist. We already had an Arc on the outer structure, and
this was what we needed to do reference-counted clones of, while the vector
structure just didn't need that. So hopefully that removes yet another level of
indirection.