Non copy shared channel (#276)
* WIP: Remove 'Copy' constraint on item type for SPSC queue

Change the Cell<T> -> UnsafeCell<Option<T>> for the SPSC queue used as the
backing ring buffer for a shared channel. This allows us to remove the Copy
constraint on T and accept a wider range of types.

* Remove alternative impl for non-copy shared channels

Remove the Copy bound on shared channels by destructuring heap allocated types,
and then storing Copy types inline. This was removed in favor of the simpler,
more versatile UnsafeCell method.

* remove unnecessary TODO comment

* Additional tests and tweaks for non-copy shared channel

Added many more tests to verify drop behavior, and attempted to simulate all
possible scenarios. Also managed to remove both the Option, and an inner Arc
that didn't need to exist. We already had an Arc on the outer structure, and
this was what we needed to do reference-counted clones of, while the vector
structure just didn't need that. So hopefully that removes yet another level of
indirection.

* remove needless match statments in shared channel tests

Didn't need match statements to check the errors in the 'drop' related shared
channel tests.