fixes #270: error information lost
Removed thiserror, and all the derive and custom attribute macros in use for our
error types. Probably has a good effect on compile times, and we were
increasingly not needing the features it provided. Added a few tests to make
sure we are printing the correct information in both display and debug printing.