Error lost context (#282)
* WIP: fix error messages data loss

Removes thiserror b/c we are forced to implement the display impl by hand,
leaving the crate pointless for our purposes. Our needs were just too
complicated, for this crate to be used. Instead we can now actually give custom
messages for WouldBlock and Closed messages based on the resource type. Because
some include space for extra fields, or messages.

* fixes #270: error information lost

Removed thiserror, and all the derive and custom attribute macros in use for our
error types. Probably has a good effect on compile times, and we were
increasingly not needing the features it provided. Added a few tests to make
sure we are printing the correct information in both display and debug printing.