return a GlommioError for rename, remove
remove and rename are done through system calls, which are converted to
io::Errors. They are then converted through GlommioErrors through
enhance, then back to io::Error for the return value.

This has no other reason except for the fact that we forgot to convert
this to GlommioError.

Aside from the needless conversions, this causes the information about
which OS error generated this to be essentially lost. It is there, but
it is behind a Custom type, that needs a lot of downcast magic to be
fetched.

This patch returns a GlommioError, and adds a convenience method to
grab the OS error that generated it when it is available.

It also adds a test to make sure we won't break this in the future.