fix asynchronous close-on-drop
Asynchronous close-on-drop is not working. Unfortunately this is not
easy to spot, until your long-running program runs out of disk space
as the files are not properly closed.

The reason is, we close through the normal Source path. But the Source
dies right after, which triggers a cancellation. What we need to do in
this case is have a special path where we don't depend on a Source.