Ensure that linked SQEs are submitted together
Linked SQEs need to occupy contiguous spots in the io_uring
submission queue, or there will be bugs (e.g. a connect_timeout call
that never times out). We use the iou prepare_sqes API to ensure this
is the case.

If the requested number of SQEs can't be filled right away, we return
to try again later. If the link chain points outside of the queue bounds
(i.e. the last queue element's link flags are set) we panic, because
this is likely an internal source prep bug.