Implement TcpStream::connect_timeout
We're using the new SQE submission flags to implement
TcpStream::connect_timeout with a connect SQE soft-linked to a
link_timeout SQE. If the timeout fires, the connect event fails with
ECANCELED. To match the standard library, we map that error to
`TimedOut` at the connect_timeout level.

The CQE processing loop was changed to no longer filter out
cancelled events because that caused connect_timeout to await on
a cancellation error that was discarded and would never arrive.

Some sys::Reactor event prep methods were also changed to avoid
double mutable borrow runtime errors in queue_request_into_ring.