add a generic way to issue timeouts
As helpful as io_uring linked SQEs are for specific operations, there
are times where you have a complex set of operations - that potentially
can spawn many sub-tasks, and need a timeout on the whole thing.

async_std has a generic "timeout" function and I think that is a good
idea. However, I am making this part of the timer, not I/O module.

The implementation is also pretty similar.