fix prepare_sqes
This method takes a slice from the sqe queue, but that is completely
wrong: the sqe queue is a circular buffer, so if you try to get 2
elements at position n - 1 you access invalid memory.

This patch transforms the very helpful SQEs structure so that it now
only has the information needed to access each SQE and constructs them
from there.