task: allow remote wakers.
Glommio is fully a thread per core library, but sometimes there is still
the need to wake up wakers remotely. One example of where this is
necessary is if we are using a third party channel library that is
designed for other executors (like Tokio) that won't uphold the thread
per core restriction.

Channels are special, because we can add one end to glommio and the
other end to any other thread running any other executor (or even
synchronous endpoint).

Since the reference counting is now atomic, we just need to enhance the
executor to be able to receive foreign wakers. We do that by using a
mpsc queue: our executor is the only consumer so that works. For a
single channel there will also only be a single producer, but we may
have multiple channels at any given time.