clean up state machine further
Jiang reports issue with reference counting that is most certainly a
leak. A lot of the state machine is more complex than it should be, in
anticipation of state being manipulated remotely, which we don't do.

This patch simplifies the reference counting by a lot, and as a result
hopefully do away with our issues.