add a bulk io submission facility
Introduces a stream interface that, given a list of sources, will yield
them once they are ready and in the order they were provided. The stream
allows the producer to submit items without a source to indicate that
this IO is fulfilled by the previous source. This will allows IO
deduplication in a later commit.

The benefit of this approach is that it makes it possible to submit an
arbitrary number of concurrent IO and wait for them from a single task,
reducing the CPU overhead. This effectively means that a Task is no
longer needed as a unit of parallelization when doing IO in glommio.