export close_rc to facilitate closing an Rc<DmaFile>
We have an internal close_rc function that we use when closing a file
wrapped in an Rc. The following commits will introduce methods that
expect a DmaFile to be wrapped in an Rc. Therefore, we need to make it
easy for applications to close those files.

This is not ideal but will do for now. An alternative solution would be
to always use an Rc inside the file and make this transparent for the
user, at the cost of paying for the Rc even when it isn't needed.