add CoalescedReads iterator to merge and deduplicate IO requests
Introduce CoalescedReads, an iterator that takes as input iterator of
IO vectors (pos, size) and outputs an iterator of deduplicated
IO vectors. The output format contains two IO vectors: the original
vector and the coalesced vector the original request was merged in.

Two user-provided arguments control the aggressiveness of the
coalescing:

* The first is the maximum size in bytes of a merged IO request. For
instance, if the user allows 16k, then the stream will never merge  IO
requests of a greater size. The stream will, however, pass through any
IO vectors with a size greater than this value.

* The second argument is optional and prescribes what read amplification
is acceptable when merging IO requests. Note that the iterator processes
inputs in order so if two subsequent IO vectors have a gap greater than
this, they will not be merged, even if a third IO vector that fits in
between will follow later on in the stream.

Example (max_merged_buffer_size: 4096, no max read amplification):

Input: (64, 256), (32, 4064), (4095, 10), (8192, 32768)
Output:
	((64, 256), (32, 4073)),
	((32, 4064), (32, 4073)),
	((4095, 10), (32, 4073)),
	((8192, 32768), (8192, 32768))

In this example, a smart reactor (wink, wink!!) could perform two IO
requests for (32, 4073) and (8192, 32768) and slice the resulting
buffers according to the user requirements.